+++ CHANGELOG.md
@@ -8,5 +8,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] - 2018-11-22
 ### Added
+- `SQLiteDatabase.*` constants
 
 [0.1.0]:  https://github.com/drydart/flutter_sqlcipher/compare/0.0.6...0.1.0